reorganise files (impl closer to their structs)
